+++ src/find.rs
@@ -1,4 +1,4 @@
-use failure::Fail;
+use thiserror::Error;
 
 use std::rc::Rc;
 
@@ -13,26 +13,20 @@ use crate::PlaybackStatus;
 const LIST_NAMES_TIMEOUT_MS: i32 = 500;
 
 /// This enum encodes possible error cases that could happen when finding players.
-#[derive(Fail, Debug)]
+#[derive(Debug, Error)]
 pub enum FindingError {
     /// No player was found matching the requirements of the calling method.
-    #[fail(display = "No player found")]
+    #[error("No player found")]
     NoPlayerFound,
 
     /// Finding failed due to an underlying [`DBusError`].
-    #[fail(display = "{}", _0)]
-    DBusError(#[cause] DBusError),
+    #[error("{}", 0)]
+    DBusError(#[from] DBusError),
 }
 
 impl From<dbus::Error> for FindingError {
     fn from(error: dbus::Error) -> Self {
         FindingError::DBusError(error.into())
-    }
-}
-
-impl From<DBusError> for FindingError {
-    fn from(error: DBusError) -> Self {
-        FindingError::DBusError(error)
     }
 }